refactored code in game.jsx file
